requirements.txt
appdirs==1.4.0
asgi-redis==1.0.0
asgiref==1.0.0
attrs==16.3.0
autobahn==0.17.1
Automat==0.5.0
channels==1.0.3
constantly==15.1.0
daphne==1.0.3
Django==1.9
django-crispy-forms==1.6.1
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
gunicorn==19.6.0
incremental==16.10.1
Markdown==2.6.8
msgpack-python==0.4.8
packaging==16.8
PyJWT==1.4.2
pyparsing==2.1.10
redis==2.10.5
six==1.10.0
Twisted==17.1.0
txaio==2.6.1
zope.interface==4.3.3
